% Generated by roxygen2 (4.0.0): do not edit by hand
\name{browsable}
\alias{browsable}
\alias{is.browsable}
\title{Make an HTML object browsable}
\usage{
browsable(x, value = TRUE)

is.browsable(x)
}
\arguments{
\item{x}{The object to make browsable or not.}

\item{value}{Whether the object should be considered browsable.}
}
\value{
\code{browsable} returns \code{x} with an extra attribute to indicate
  that the value is browsable.

\code{is.browsable} returns \code{TRUE} if the value is browsable, or
  \code{FALSE} if not.
}
\description{
By default, HTML objects display their HTML markup at the console when
printed. \code{browsable} can be used to make specific objects render as HTML
by default when printed at the console.
}
\details{
You can override the default browsability of an HTML object by explicitly
passing \code{browse = TRUE} (or \code{FALSE}) to the \code{print} function.
}

