% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_print.R
\name{console_print}
\alias{console_print}
\title{Print the HTML in the console}
\usage{
console_print(html, ..., background = "white", viewer = get_console_viewer())
}
\arguments{
\item{html}{HTML content to print}

\item{...}{Additional arguments for \code{viewer}}

\item{background}{Background color for web page}

\item{viewer}{A function to be called with the URL or path to the generated
HTML page. Can be \code{NULL}, in which case no viewer will be invoked.}
}
\value{
Invisibly returns the URL or path of the generated HTML page.
}
\description{
Print the HTML in the console
}
