% Generated by roxygen2 (4.0.0): do not edit by hand
\name{copy_deps}
\alias{copy_deps}
\title{Copy dependencies to directory}
\usage{
copy_deps(dependencies, src = ".", dest = NULL)
}
\arguments{
\item{dependencies}{List of \code{\link{html_dependency}}s.}

\item{src}{Base directory to use for location of files.}

\item{dest}{Destination directory}
}
\description{
Copy dependencies to directory
}
\examples{
\dontrun{
d3 <- html_dependency(name = "d3", version = "3.4.1", path = "lib/d3",
  script = "d3.min.js")
copy_deps(list(d3), system.file("www", package = "ggvis"), "~/Desktop")
}
}

