% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{HTML}
\alias{HTML}
\title{Mark Characters as HTML}
\usage{
HTML(..., .noWS = NULL)
}
\arguments{
\item{...}{Text to mark with HTML. Any additional values after the first will
to be converted to character, and all will be concatenated together.}

\item{.noWS}{Character vector used to omit some of the whitespace that would
normally be written around this HTML. Valid options include \code{before},
\code{after}, and \code{outside} (equivalent to \code{before} and
\code{end}).}
}
\value{
The input \code{text}, but marked as HTML.
}
\description{
Marks the given text as HTML, which means the \link{tag} functions will know
not to perform HTML escaping on it.
}
\examples{
el <- div(HTML("I like <u>turtles</u>"))
cat(as.character(el))

}
