% Generated by roxygen2 (4.0.1): do not edit by hand
\name{htmlDependencies}
\alias{attachDependencies}
\alias{htmlDependencies}
\alias{htmlDependencies<-}
\title{HTML dependency metadata}
\usage{
htmlDependencies(x)

htmlDependencies(x) <- value

attachDependencies(x, value)
}
\arguments{
\item{x}{An object which has (or should have) HTML dependencies.}

\item{value}{An HTML dependency, or a list of HTML dependencies.}
}
\description{
Gets or sets the HTML dependencies associated with an object (such as a tag).
}
\details{
\code{attachDependencies} provides an alternate syntax for setting
dependencies and is essentially equivalent to
\code{local(\{htmlDependencies(x) <- value; x\})}.
}

