% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{knitr_methods}
\alias{knitr_methods}
\alias{knit_print.shiny.tag}
\alias{knit_print.html}
\alias{knit_print.shiny.tag.list}
\alias{knit_print.html_dependency}
\title{Knitr S3 methods}
\usage{
knit_print.shiny.tag(x, ..., inline = FALSE)

knit_print.html(x, ..., inline = FALSE)

knit_print.shiny.tag.list(x, ..., inline = FALSE)

knit_print.html_dependency(x, ..., inline = FALSE)
}
\arguments{
\item{x}{Object to knit_print}

\item{...}{Additional knit_print arguments}

\item{inline}{Whether or not the code chunk is inline.}
}
\description{
These S3 methods are necessary to allow HTML tags to print themselves in
knitr/rmarkdown documents.
}
