% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{parseCssColors}
\alias{parseCssColors}
\title{Parse CSS color strings}
\usage{
parseCssColors(str, mustWork = TRUE)
}
\arguments{
\item{str}{CSS color strings}

\item{mustWork}{If true, invalid color strings will cause an error; if false,
then the result will contain `NA` for invalid colors.}
}
\description{
Parses/normalizes CSS color strings, and returns them as strings in
`"#RRGGBB"` and/or `"#RRGGBBAA"` format. Understands hex colors in 3, 4, 6,
and 8 digit forms, `rgb()`/`rgba()`, `hsl()`/`hsla()`, and color keywords.
}
\examples{

parseCssColors(c(
  "#0d6efd",
  "#DC35457F",
  "rgb(32,201,151)",
  "  rgba( 23 , 162 , 184 , 0.5 )  ",
  "hsl(261, 51\%, 51\%)",
  "cornflowerblue"
))
}
