% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.shiny.tag}
\alias{print.html}
\alias{print.shiny.tag}
\title{Print method for HTML/tags}
\usage{
\method{print}{shiny.tag}(x, browse = is.browsable(x), ...)

\method{print}{html}(x, ..., selfcontained = FALSE,
  browse = is.browsable(x))
}
\arguments{
\item{x}{The value to print.}

\item{browse}{If \code{TRUE}, the HTML will be rendered and displayed in a
browser (or possibly another HTML viewer supplied by the environment via
the \code{viewer} option). If \code{FALSE} then the HTML object's markup
will be rendered at the console.}

\item{...}{Additional arguments passed to print.}
}
\description{
S3 method for printing HTML that prints markup or renders HTML in a web
browser.
}

