% Generated by roxygen2 (4.0.0): do not edit by hand
\name{renderDependencies}
\alias{renderDependencies}
\title{Create HTML for dependencies}
\usage{
renderDependencies(dependencies, srcType = c("file", "href"),
  encodeFunc = urlEncodePath, hrefFilter = identity)
}
\arguments{
\item{dependencies}{A list of \code{htmlDependency} objects.}

\item{srcType}{The type of src paths to use; valid values are \code{file} or
\code{href}.}

\item{encodeFunc}{The function to use to encode the path part of a URL. The
default should generally be used.}

\item{hrefFilter}{A function used to transform the final, encoded URLs of
  script and stylsheet files. The default should generally be used.}
}
\description{
Create the appropriate HTML markup for including these dependencies in an
HTML document.
}

