% Generated by roxygen2 (4.0.0): do not edit by hand
\name{doRenderTags}
\alias{doRenderTags}
\alias{renderTags}
\title{Render tags into HTML}
\usage{
doRenderTags(ui, indent = 0)

renderTags(ui, singletons = character(0), indent = 0)
}
\arguments{
\item{ui}{Tag object(s) to render}

\item{singletons}{A list of \link{singleton} signatures to consider already
rendered; any matching singletons will be dropped instead of rendered.}

\item{indent}{Initial indent level, or \code{FALSE} if no indentation should
  be used.}
}
\value{
Returns a list with the following variables:
\describe{
  \item{\code{head}}{An \code{\link{HTML}} string that should be included in
    \code{<head>}.
  }
  \item{\code{singletons}}{Character vector of singleton signatures that are
    known after rendering.
  }
  \item{\code{dependencies}}{A list of \link[=resolveDependencies]{resolved}
    \code{\link{htmlDependency}} objects.
  }
  \item{\code{html}}{An \code{\link{HTML}} string that represents the main
    HTML that was rendered.
  }
}
}
\description{
Renders tags and objects that can be converted into tags into HTML.
}
\details{
\code{doRenderTags} ignores singleton, head, and dependency
  handling, and simply renders the given tag objects as HTML.
}

