% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{resolveFunctionalDependencies}
\alias{resolveFunctionalDependencies}
\title{Resolve tag functions}
\usage{
resolveFunctionalDependencies(x)
}
\arguments{
\item{x}{a list of HTML dependencies or tags.}
}
\value{
a list of HTML dependencies.
}
\description{
Resolves any [tagFunction()]s inside a list of [htmlDependencies()]. To
resolve [tagFunction()]s _and then_ remove redundant dependencies all at once,
use [resolveDependencies()] (which calls this function internally).
}
\seealso{
[tagFunction()], [resolveDependencies()]
}
