% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save_html}
\alias{save_html}
\title{Save an HTML object to a file}
\usage{
save_html(html, file, libdir = "lib")
}
\arguments{
\item{html}{HTML content to print}

\item{file}{File to write content to}

\item{libdir}{Directory to copy dependenies to}
}
\description{
Save the specified HTML object to a file, copying all of it's
dependencies to the directory specified via \code{libdir}.
}

