% Generated by roxygen2 (4.0.0): do not edit by hand
\name{surroundSingletons}
\alias{surroundSingletons}
\alias{takeSingletons}
\title{Singleton manipulation functions}
\usage{
surroundSingletons(ui)

takeSingletons(ui, singletons = character(0), desingleton = TRUE)
}
\description{
Singleton manipulation functions
}

