% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.R
\name{as.tags}
\alias{as.tags}
\title{Convert a value to tags}
\usage{
as.tags(x, ...)
}
\arguments{
\item{x}{Object to be converted.}

\item{...}{Any additional parameters.}
}
\description{
An S3 method for converting arbitrary values to a value that can be used as
the child of a tag or \code{tagList}. The default implementation simply calls
\code{\link[base]{as.character}}.
}

