% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findDependencies}
\alias{findDependencies}
\title{Collect attached dependencies from HTML tag object}
\usage{
findDependencies(ui)
}
\arguments{
\item{ui}{A tag-like object to search for dependencies.}
}
\value{
A list of \code{\link{htmlDependency}} objects.
}
\description{
Walks a hierarchy of tags looking for attached dependencies.
}

