% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{findDependencies}
\alias{findDependencies}
\title{Collect attached dependencies from HTML tag object}
\usage{
findDependencies(tags, tagify = TRUE)
}
\arguments{
\item{tags}{A tag-like object to search for dependencies.}

\item{tagify}{Whether to tagify the input before searching for dependencies.}
}
\value{
A list of \code{\link{htmlDependency}} objects.
}
\description{
Walks a hierarchy of tags looking for attached dependencies.
}

