% Generated by roxygen2 (4.0.1): do not edit by hand
\name{html_print}
\alias{html_print}
\title{Implementation of the print method for HTML}
\usage{
html_print(html, background = "white")
}
\arguments{
\item{html}{HTML content to print}

\item{background}{Background color for web page}
}
\description{
Convenience method that provides an implementation of the
\code{\link[base:print]{print}} method for HTML content.
}

