% Generated by roxygen2 (4.0.0): do not edit by hand
\name{html_print}
\alias{html_print}
\title{Implementation of the print method for HTML}
\usage{
html_print(html)
}
\arguments{
\item{html}{HTML content to print}

\item{dependencies}{List of HTML dependencies created using
the \code{\link{htmlDependency}} function.}
}
\description{
Convenience method that provides an implementation of the
\code{\link[base:print]{print}} method for HTML content.
}

