% Generated by roxygen2 (4.0.2): do not edit by hand
\name{htmlDependencies}
\alias{attachDependencies}
\alias{htmlDependencies}
\alias{htmlDependencies<-}
\title{HTML dependency metadata}
\usage{
htmlDependencies(x)

htmlDependencies(x) <- value

attachDependencies(x, value, append = FALSE)
}
\arguments{
\item{x}{An object which has (or should have) HTML dependencies.}

\item{value}{An HTML dependency, or a list of HTML dependencies.}

\item{append}{If FALSE (the default), replace any existing dependencies. If
  TRUE, add the new dependencies to the existing ones.}
}
\description{
Gets or sets the HTML dependencies associated with an object (such as a tag).
}
\details{
\code{attachDependencies} provides an alternate syntax for setting
dependencies. It is similar to \code{local(\{htmlDependencies(x) <- value;
x\})}, except that if there are any existing dependencies,
\code{attachDependencies} will add to them, instead of replacing them.
}

