% Generated by roxygen2 (4.0.0): do not edit by hand
\name{htmlDependency}
\alias{htmlDependency}
\title{Define an HTML dependency}
\usage{
htmlDependency(name, version, src, meta = NULL, script = NULL,
  stylesheet = NULL, head = NULL)
}
\arguments{
\item{name}{Library name}

\item{version}{Library version}

\item{src}{Unnamed single-element character vector indicating the full path
of the library directory. Alternatively, a named character string with one
or more elements, indicating different places to find the library; see
Details.}

\item{meta}{Named list of meta tags to insert into document head}

\item{script}{Script(s) to include within the document head (should be
specified relative to the \code{path} parameter).}

\item{stylesheet}{Stylesheet(s) to include within the document (should be
specified relative to the \code{path} parameter).}

\item{head}{Arbitrary lines of HTML to insert into the document head}
}
\value{
An object that can be included in the list of dependencies passed to
  \code{\link{html_print}} or \code{\link{html_knit_print}}.
}
\description{
Define an HTML dependency (e.g. CSS or Javascript and related library). HTML
dependency definitions are required for \code{\link{html_output}} that
require CSS or JavaScript within the document head to render correctly.
}
\details{
Each dependency can be located on the filesystem, at a relative or
  absolute URL, or both. The location types are indicated using the names of
  the \code{src} character vector: \code{file} for filesystem directory,
  \code{href} for URL. For example, a dependency that was both on disk and
  at a URL might use \code{src = c(file=filepath, href=url)}.

  See the documentation on
  \href{http://rmarkdown.rstudio.com/developer_html_widgets.html}{R Markdown
  HTML Widgets} for examples and additional details.
}

