% Generated by roxygen2 (4.0.0): do not edit by hand
\name{htmlEscape}
\alias{htmlEscape}
\title{Escape HTML entities}
\usage{
htmlEscape(text, attribute = FALSE)
}
\arguments{
\item{text}{Text to escape}

\item{attribute}{Escape for use as an attribute value}
}
\value{
Character vector with escaped text.
}
\description{
Escape HTML entities contained in a character vector so that it can be safely
included as text or an attribute value within an HTML document
}

