% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools-package.R
\docType{package}
\name{htmltools-package}
\alias{htmltools}
\alias{htmltools-package}
\title{htmltools: Tools for HTML}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for HTML generation and output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/htmltools}
  \item \url{https://rstudio.github.io/htmltools/}
  \item Report bugs at \url{https://github.com/rstudio/htmltools/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Yihui Xie \email{yihui@posit.co}
  \item Jeff Allen
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
