% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{include}
\alias{include}
\alias{includeHTML}
\alias{includeHTMLDocument}
\alias{includeText}
\alias{includeMarkdown}
\alias{includeCSS}
\alias{includeScript}
\title{Include Content From a File}
\usage{
includeHTML(path)

includeHTMLDocument(path, ...)

includeText(path)

includeMarkdown(path)

includeCSS(path, ...)

includeScript(path, ...)
}
\arguments{
\item{path}{The path of the file to be included. It is highly recommended to
use a relative path (the base path being the Shiny application directory),
not an absolute path.}

\item{...}{Any additional attributes to be applied to the generated tag.}
}
\description{
Load HTML, text, or rendered Markdown from a file and turn into HTML.
}
\details{
These functions provide a convenient way to include an extensive amount of
HTML, textual, Markdown, CSS, or JavaScript content, rather than using a
large literal R string.
}
\note{
\code{includeHTML()} directly inlines the contents of the file, which is only
suitable for fragments of HTML. To include a full HTML document, such as
the generated output of an HTML R Markdown or Quarto document, use
\code{includeHTMLDocument()}.

\code{includeHTMLDocument()} embeds the contents of the file in an
\verb{<iframe>} element. When \code{path} is detected to be a full URL starting with
\verb{https://}, the \code{src} attribute of the \verb{<iframe>} will be set to the URL.
When \code{path} is a local file, the contents of the file will be embedded in
the \verb{<iframe>} using the \code{srcdoc} attribute, which embeds the complete HTML
document directly in the page where \code{includeHTMLDocument()} was called.

\code{includeText} escapes its contents, but does no other processing.
This means that hard breaks and multiple spaces will be rendered as they
usually are in HTML: as a single space character. If you are looking for
preformatted text, wrap the call with \code{\link[=pre]{pre()}}, or consider using
\code{includeMarkdown} instead.

The \code{includeMarkdown} function requires the \code{markdown}
package.
}
