% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency.R
\name{renderDependencies}
\alias{renderDependencies}
\title{Create HTML for dependencies}
\usage{
renderDependencies(
  dependencies,
  srcType = c("href", "file"),
  encodeFunc = urlEncodePath,
  hrefFilter = identity
)
}
\arguments{
\item{dependencies}{A list of \code{htmlDependency} objects.}

\item{srcType}{The type of src paths to use; valid values are \code{file} or
\code{href}.}

\item{encodeFunc}{The function to use to encode the path part of a URL. The
default should generally be used.}

\item{hrefFilter}{A function used to transform the final, encoded URLs of
script and stylesheet files. The default should generally be used.}
}
\value{
An \code{\link{HTML}} object suitable for inclusion in the head of an
  HTML document.
}
\description{
Create the appropriate HTML markup for including dependencies in an HTML
document.
}
