% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{resolveDependencies}
\alias{resolveDependencies}
\title{Resolve a list of dependencies}
\usage{
resolveDependencies(dependencies, resolvePackageDir = TRUE)
}
\arguments{
\item{dependencies}{A list of \code{\link{htmlDependency}} objects.}

\item{resolvePackageDir}{Whether to resolve the relative path to an absolute
path via \code{\link{system.file}} when the \code{package} attribute is
present in a dependency object.}
}
\value{
dependencies A list of \code{\link{htmlDependency}} objects with
  redundancies removed.
}
\description{
Given a list of dependencies, removes any redundant dependencies (based on
name equality). If multiple versions of a dependency are found, the copy with
the latest version number is used.
}
