% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{resolveFunctionalDependencies}
\alias{resolveFunctionalDependencies}
\title{Resolve functional HTML dependencies}
\usage{
resolveFunctionalDependencies(dependencies)
}
\arguments{
\item{dependencies}{A list of \code{\link{htmlDependency}} objects.}
}
\description{
Resolves any [tagFunction()]s inside a list of [htmlDependencies()]. To
resolve [tagFunction()]s _and then_ remove redundant dependencies all at once,
use [resolveDependencies()] (which calls this function internally).
}
\seealso{
[tagFunction()], [resolveDependencies()]
}
