% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_print.R
\name{save_html}
\alias{save_html}
\alias{save_html.default}
\title{Save an HTML object to a file}
\usage{
save_html(html, file, ...)

\method{save_html}{default}(html, file, background = "white", libdir = "lib", lang = "en", ...)
}
\arguments{
\item{html}{HTML content to print}

\item{file}{File path or connection. If a file path containing a
sub-directory, the sub-directory must already exist.}

\item{background}{Background color for web page}

\item{libdir}{Directory to copy dependencies to}

\item{lang}{Value of the \verb{<html>} \code{lang} attribute}
}
\description{
An S3 generic method for saving an HTML-like object to a file. The default
method copies dependency files to the directory specified via \code{libdir}.
}
