% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tagList}
\alias{tagList}
\title{Create a list of tags}
\usage{
tagList(...)
}
\arguments{
\item{...}{A collection of \link{tag}s.}
}
\description{
Create a \code{list()} of \link{tag}s with methods for \code{\link[=print]{print()}}, \code{\link[=as.character]{as.character()}},
etc.
}
\examples{
tagList(
  h1("Title"),
  h2("Header text", style = "color: red;"),
  p("Text here")
)
}
